ingestors/APIcopy/requirements.txt
flask 
graphene 
ariadne
graphql-server[flask]
pymongo
pydantic
python-dateutil
rx
influxdb
flask-socketio
django
graphene-django
channels
fastapi
ariadne
graphql-ws
